+++ src/OpenTelemetry.Instrumentation.AspNetCore/Implementation/HttpInMetricsListener.cs
@@ -24,13 +24,11 @@ using System.Diagnostics.CodeAnalysis;
 using Microsoft.AspNetCore.Routing;
 #endif
 using OpenTelemetry.Trace;
-using static OpenTelemetry.Internal.HttpSemanticConventionHelper;
 
 namespace OpenTelemetry.Instrumentation.AspNetCore.Implementation;
 
 internal sealed class HttpInMetricsListener : ListenerHandler
 {
-    internal const string HttpServerDurationMetricName = "http.server.duration";
     internal const string HttpServerRequestDurationMetricName = "http.server.request.duration";
 
     internal const string OnUnhandledHostingExceptionEvent = "Microsoft.AspNetCore.Hosting.UnhandledException";
@@ -43,31 +41,13 @@ internal sealed class HttpInMetricsListener : ListenerHandler
     private static readonly object ErrorTypeHttpContextItemsKey = new();
 
     private readonly Meter meter;
-    private readonly AspNetCoreMetricsInstrumentationOptions options;
-    private readonly Histogram<double> httpServerDuration;
     private readonly Histogram<double> httpServerRequestDuration;
-    private readonly bool emitOldAttributes;
-    private readonly bool emitNewAttributes;
 
-    internal HttpInMetricsListener(string name, Meter meter, AspNetCoreMetricsInstrumentationOptions options)
+    internal HttpInMetricsListener(string name, Meter meter)
         : base(name)
     {
         this.meter = meter;
-        this.options = options;
-
-        this.emitOldAttributes = this.options.HttpSemanticConvention.HasFlag(HttpSemanticConvention.Old);
-
-        this.emitNewAttributes = this.options.HttpSemanticConvention.HasFlag(HttpSemanticConvention.New);
-
-        if (this.emitOldAttributes)
-        {
-            this.httpServerDuration = meter.CreateHistogram<double>(HttpServerDurationMetricName, "ms", "Measures the duration of inbound HTTP requests.");
-        }
-
-        if (this.emitNewAttributes)
-        {
-            this.httpServerRequestDuration = meter.CreateHistogram<double>(HttpServerRequestDurationMetricName, "s", "Duration of HTTP server requests.");
-        }
+        this.httpServerRequestDuration = meter.CreateHistogram<double>(HttpServerRequestDurationMetricName, "s", "Duration of HTTP server requests.");
     }
 
     public override void OnEventWritten(string name, object payload)
@@ -77,24 +57,13 @@ internal sealed class HttpInMetricsListener : ListenerHandler
             case OnUnhandledDiagnosticsExceptionEvent:
             case OnUnhandledHostingExceptionEvent:
                 {
-                    if (this.emitNewAttributes)
-                    {
-                        this.OnExceptionEventWritten(name, payload);
-                    }
+                    this.OnExceptionEventWritten(name, payload);
                 }
 
                 break;
             case OnStopEvent:
                 {
-                    if (this.emitOldAttributes)
-                    {
-                        this.OnEventWritten_Old(name, payload);
-                    }
-
-                    if (this.emitNewAttributes)
-                    {
-                        this.OnEventWritten_New(name, payload);
-                    }
+                    this.OnStopEventWritten(name, payload);
                 }
 
                 break;
@@ -106,7 +75,7 @@ internal sealed class HttpInMetricsListener : ListenerHandler
         // We need to use reflection here as the payload type is not a defined public type.
         if (!TryFetchException(payload, out Exception exc) || !TryFetchHttpContext(payload, out HttpContext ctx))
         {
-            AspNetCoreInstrumentationEventSource.Log.NullPayload(nameof(HttpInMetricsListener), nameof(this.OnExceptionEventWritten), HttpServerDurationMetricName);
+            AspNetCoreInstrumentationEventSource.Log.NullPayload(nameof(HttpInMetricsListener), nameof(this.OnExceptionEventWritten), HttpServerRequestDurationMetricName);
             return;
         }
 
@@ -127,56 +96,7 @@ internal sealed class HttpInMetricsListener : ListenerHandler
             => HttpContextPropertyFetcher.TryFetch(payload, out ctx) && ctx != null;
     }
 
-    public void OnEventWritten_Old(string name, object payload)
-    {
-        var context = payload as HttpContext;
-
-        if (context == null)
-        {
-            AspNetCoreInstrumentationEventSource.Log.NullPayload(nameof(HttpInMetricsListener), EventName, HttpServerDurationMetricName);
-            return;
-        }
-
-        // TODO: Prometheus pulls metrics by invoking the /metrics endpoint. Decide if it makes sense to suppress this.
-        // Below is just a temporary way of achieving this suppression for metrics (we should consider suppressing traces too).
-        // If we want to suppress activity from Prometheus then we should use SuppressInstrumentationScope.
-        if (context.Request.Path.HasValue && context.Request.Path.Value.Contains("metrics"))
-        {
-            return;
-        }
-
-        TagList tags = default;
-
-        tags.Add(new KeyValuePair<string, object>(SemanticConventions.AttributeHttpFlavor, HttpTagHelper.GetFlavorTagValueFromProtocol(context.Request.Protocol)));
-        tags.Add(new KeyValuePair<string, object>(SemanticConventions.AttributeHttpScheme, context.Request.Scheme));
-        tags.Add(new KeyValuePair<string, object>(SemanticConventions.AttributeHttpMethod, context.Request.Method));
-        tags.Add(new KeyValuePair<string, object>(SemanticConventions.AttributeHttpStatusCode, TelemetryHelper.GetBoxedStatusCode(context.Response.StatusCode)));
-
-        if (context.Request.Host.HasValue)
-        {
-            tags.Add(new KeyValuePair<string, object>(SemanticConventions.AttributeNetHostName, context.Request.Host.Host));
-
-            if (context.Request.Host.Port is not null && context.Request.Host.Port != 80 && context.Request.Host.Port != 443)
-            {
-                tags.Add(new KeyValuePair<string, object>(SemanticConventions.AttributeNetHostPort, context.Request.Host.Port));
-            }
-        }
-
-#if NET6_0_OR_GREATER
-        var route = (context.GetEndpoint() as RouteEndpoint)?.RoutePattern.RawText;
-        if (!string.IsNullOrEmpty(route))
-        {
-            tags.Add(new KeyValuePair<string, object>(SemanticConventions.AttributeHttpRoute, route));
-        }
-#endif
-
-        // We are relying here on ASP.NET Core to set duration before writing the stop event.
-        // https://github.com/dotnet/aspnetcore/blob/d6fa351048617ae1c8b47493ba1abbe94c3a24cf/src/Hosting/Hosting/src/Internal/HostingApplicationDiagnostics.cs#L449
-        // TODO: Follow up with .NET team if we can continue to rely on this behavior.
-        this.httpServerDuration.Record(Activity.Current.Duration.TotalMilliseconds, tags);
-    }
-
-    public void OnEventWritten_New(string name, object payload)
+    public void OnStopEventWritten(string name, object payload)
     {
         var context = payload as HttpContext;
         if (context == null)